Hoist Client classes into gro package and update samples to use gro package
