+++ api/README.md
@@ -1,0 +1,5 @@
+# Deprecation warning
+
+The package in this subdirectory is deprecated and will be removed in a future release.
+
+Please import from `gro` ([/gro](/gro)) and do not import `api.client`.
