+++ samples/similar_regions/similar_region.py
@@ -1,7 +1,7 @@
 import numpy as np
 import os
 from sklearn.neighbors import BallTree
-from api.client.batch_client import BatchClient
+from gro import BatchClient
 from api.client.lib import get_default_logger
 from similar_region_state import SimilarRegionState
 from sklearn.metrics.pairwise import euclidean_distances
